+++ lib/src/analyzers/unnecessary_nullable_analyzer/unnecessary_nullable_analysis_config.dart
@@ -1,0 +1,12 @@
+import 'package:glob/glob.dart';
+
+/// Represents converted unused code config which contains parsed entities.
+class UnnecessaryNullableAnalysisConfig {
+  final Iterable<Glob> globalExcludes;
+  final Iterable<Glob> analyzerExcludedPatterns;
+
+  const UnnecessaryNullableAnalysisConfig(
+    this.globalExcludes,
+    this.analyzerExcludedPatterns,
+  );
+}